Complete OBS service refactoring and bug fixes.
The refactored architecture now properly separates concerns across
9 focused modules while maintaining full backward compatibility.
